remove some dist uts
