+++ tools/parallel_UT_rule.py
@@ -486,12 +486,9 @@ HIGH_PARALLEL_JOB_NEW = [
     'test_dist_fleet_decay',
     'test_dist_fleet_simnet',
     'test_dist_sparse_load_ps1',
-    'test_dist_mnist_fleet_save',
     'test_dist_fleet_ps7',
-    'test_dist_mnist_fleetapi',
     'test_dist_sparse_tensor_load_adam',
     'test_dist_fleet_ps_gpu_ctr',
-    'test_dist_mnist_ring_allreduce',
     'test_dist_op',
     'test_new_group_api',
     'test_dist_fleet_heter_base',
@@ -506,10 +503,8 @@ HIGH_PARALLEL_JOB_NEW = [
     'test_dist_fleet_a_sync_optimizer_auto_async',
     'test_dist_fleet_a_sync_optimizer_auto',
     'test_dist_fleet_ps9',
-    'test_dist_fleet_raw_program_optimizer_fuse_allreduce',
     'test_dist_fleet_ps11',
     'test_dist_fleet_ps8',
-    'test_dist_mnist_fp16_allreduce',
     'test_dist_fleet_ps12',
     'test_collective_split_row_linear',
     'test_collective_reduce_api',
@@ -520,19 +515,14 @@ HIGH_PARALLEL_JOB_NEW = [
     'test_collective_split_embedding_none_divisible',
     'test_parallel_dygraph_dataparallel',
     'test_dist_fleet_ps3',
-    'test_dist_mnist_pg',
     'test_pipeline_parallel',
     'test_fleet_perf_test',
     'test_dist_fleet_ps5',
     'test_dist_fleet_sparse_embedding_ctr',
     'test_collective_broadcast_api',
     'retry_allocator_test',
-    'test_dist_mnist_backward_deps',
-    'test_dist_mnist_multi_comm',
-    'test_dist_allreduce_op',
     'test_parallel_dygraph_sparse_embedding',
     'test_dist_se_resnext_dgc',
-    'test_dist_sharding_save',
     'test_dist_fleet_a_sync_optimizer_async',
     'test_gen_nccl_id_op',
     'test_auto_checkpoint',
@@ -540,12 +530,10 @@ HIGH_PARALLEL_JOB_NEW = [
     'test_parallel_dygraph_sparse_embedding_over_height',
     'test_dist_fleet_ps4',
     'test_collective_alltoall_api',
-    'test_dist_fleet_raw_program_optimizer',
     'test_parallel_dygraph_mp_layers',
     'test_dist_fleet_geo',
     'test_fleet_raw_program_meta_optimizer',
     'test_sync_batch_norm_op',
-    'test_dist_mnist_batch_merge',
     'test_fleet_launch_ps',
     'test_dist_sparse_tensor_load_sgd_deprecated',
     'test_dist_fleet_a_sync_optimizer_auto_geo',
@@ -561,7 +549,6 @@ HIGH_PARALLEL_JOB_NEW = [
     'test_new_group',
     'test_imperative_signal_handler',
     'test_parallel_dygraph_sharding_parallel',
-    'test_dist_hapi_model',
     'test_dist_mnist_gradient_merge',
     'test_ir_pass_pipeline',
     'test_rnn_dp',
@@ -1966,34 +1953,25 @@ CPU_PARALLEL_JOB = [
     'auto_growth_best_fit_allocator_test',
     'assign_op_test',
     'aes_cipher_test',
-    'test_dist_mnist_fp16_allreduce',
     'test_dist_mnist_gradient_merge',
-    'test_dist_allreduce_op',
     'test_parallel_dygraph_se_resnext',
     'test_dist_fleet_ps9',
     'test_dist_fleet_infer',
     'test_dist_se_resnext_sync',
     'test_dist_oneps',
     'test_dist_sparse_load_ps1',
-    'test_dist_mnist_batch_merge',
     'test_dist_fleet_ctr',
     'test_dist_fleet_ps10',
     'test_parallel_dygraph_transformer',
-    'test_dist_mnist_fleetapi',
     'test_dist_sparse_tensor_load_adam',
     'test_dist_fleet_ps4',
     'test_dist_fleet_heter_program',
     'test_parallel_dygraph_sparse_embedding_over_height',
-    'test_dist_sharding_save',
     'test_dist_fleet_ps_gpu_ctr',
-    'test_dist_mnist_backward_deps',
     'test_dist_fleet_heter_base',
     'test_dist_sparse_tensor_load_sgd_deprecated',
     'test_new_group',
-    'test_dist_mnist_with_program',
-    'test_dist_mnist_pg',
     'test_dist_fleet_ps6',
-    'test_dist_mnist_fleet_save',
     'test_dist_fleet_a_sync_optimizer_sync',
     'test_dist_fleet_ps3',
     'test_dist_se_resnext_nccl',
@@ -2010,10 +1988,8 @@ CPU_PARALLEL_JOB = [
     'test_parallel_dygraph_dataparallel',
     'test_dist_mnist_dgc_nccl',
     'test_parallel_dygraph_unused_variables',
-    'test_dist_mnist_multi_comm',
     'test_gen_nccl_id_op',
     'test_parallel_dygraph_sparse_embedding',
-    'test_dist_mnist_ring_allreduce',
     'test_fleet_launch_async',
     'test_dist_fleet_a_sync_optimizer_geo_deprecated',
     'test_auto_checkpoint',
